Add an op that reprofiles as sRGB
Sometimes designers upload images with colors encoded
as AdobeRGB. This causes the image to look fainted when
displayed in a browser, because the RGB values assume much more
intense primaries than the ones sRGB provides.

To combat that, you need to convert the values of the primaries
to their sRGB equivalents, by applying the sRGB color profile (or,
rather, by forcing a profile conversion to take place - similar to
"Convert To Profile" in Photoshop).