Add pluggable Writer objects for saving renders
Formats we use are likely to support diverging parameters. For example, a PNG
will permit to have an indexed color as well as a true color PNG, while JPEG
only supports a quality factor. PNG in ImageMagick _also_ supports a quality factor,
but does it does not have the same semantics (it alters the palette size).

So, in practice, if you want to "force format X" you also want to be able to set
parameters specific to that format. That is not possible to do since we do not
have multivariate parameters for operators.

Therefore we replace a hardcoded output routine in RenderEngine with a set of
pluggable Writer objects we can use to write out specific formats. We also
introduce a metadata Hash object that is passed to every Operator in succession.
The operator can then inject data in there for the subsequent operators to use.

We make use of this facility to configure the :writer key to output JPEG
specifically, with a predefined quality setting.

The settings for permitting/forbidding formats in RenderEngine are removed.
In practice it is unlikely we will want to serve TIFFs from
ImageVise in the near future, and the limiting of the output
formats now lives in the auto output selector anyway