Remove the fork+process feature outright
Using fork() without exec() is not safe, especially in threaded contexts.
A living proof of this is that the test suite either crashes or hangs on
Travis. For more info, see

  http://www.evanjones.ca/fork-is-dangerous.html

If we really want _isolated_ processing at some point we need to write out the
image processing script (serialize the pipeline) and exec() that script in the
subprocess, and pick up the processed image afterwards. Given that in production
RMagick proved to be memory-stable as long as we are very rigorous with calling
`Magick::Image#destroy!` on everything we allocate, the primary reason - memory
containment - is moot.

From the standpoint of security processing within an isolated process makes
all the sense, but the approach we had is not the way it has to be done.