Integrate FormatParser instead of magic_bytes
The process of format detection and the resolution
checks belong in the base image_vise library.
We can also use it to output the rendered file, and
we win by not having to check the file format post-download
if we verified it once already.

A more secure alternative would be to do the fetch _first_
probably... since then it would be much harder to exploit